refactor string manipulation into stringUtils module
