+++ frontend/types/index.ts
@@ -1,0 +1,22 @@
+export interface ApiArticle {
+    id: number;
+    title: string;
+    source: string;
+    createdAt: string;
+    updatedAt: string;
+    summary: { String: string; Valid: boolean } | null;
+    link: string;
+    is_on_homepage: boolean;
+}
+  
+export interface Article {
+    id: number;
+    title: string;
+    source: string;
+    createdAt: string;
+    updatedAt: string;
+    summary: string;
+    link: string;
+    isOnHomepage: boolean;
+}
+
